Fix/be testing (#26)
* [BE feat] Member MockBean testing 구현

* [BE feat] Member Mockbean testing 구현

* Changes by superscman

* [BE feat] Member api 자동화 구현

* [BE feat] Member Mockbean testing 구현

* [BE feat] Member api 자동화 구현

+++ server/pre_project/src/main/java/com/codestates/backend/pre_project/point/entity/Point.java
@@ -24,7 +24,7 @@ public class Point {
     private long pointId;
 
     @Column
-    private long pointCount;
+    private int pointCount;
 
     @NotNull
     @OneToOne
